Move replaceNonMatching fn
Move replaceNonMatching function from wallets/index.tsx to src/utils/replaceNonMatching.ts for better readability and infrastucture.
Add javadoc for better devex.
